+++ lib/in_app_browser.dart
@@ -1,12 +1,10 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_webview_plugin/flutter_webview_plugin.dart';
 import 'package:in_app_browser/MyManagement/HistoryStorage.dart';
 import 'package:share/share.dart';
-import 'package:sqflite/sqflite.dart';
-// import 'package:webview_flutter/webview_flutter.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 
 List<String> months = [
   "January",
@@ -94,8 +92,8 @@ class InAppBrowser extends StatefulWidget {
 class _InAppBrowserState extends State<InAppBrowser>
     with TickerProviderStateMixin {
   final flutterWebViewPlugin = FlutterWebviewPlugin();
-  Database _storeDB;
-  HistoryStoreProvider _storeProvider = HistoryStoreProvider();
+
+  SharedPreferences prefs;
 
   List<Map<dynamic, dynamic>> _myList = [];
   bool isLoading = true;
@@ -124,20 +122,30 @@ class _InAppBrowserState extends State<InAppBrowser>
 
     Future.delayed(Duration.zero, () {
       isLoading = false;
-      initialDB();
-      setState(() {});
+      initailSP();
     });
   }
 
-  void initialDB() async {
-    _storeDB = await _storeProvider.open();
+  void initailSP() async {
+    prefs = await SharedPreferences.getInstance();
+    List<String> mKeys = prefs.getKeys().toList();
 
-    _storeProvider.getAllHistoryItem().then((value) {
-      if (value != null)
-        value.forEach((element) {
-          _myList.add(element);
-        });
+    if (mKeys.isNotEmpty)
+      mKeys.forEach((key) {
+        if (key.contains("URL: ")) {
+          String value = prefs.getString(key);
+          _myList.add({
+            "title": value.substring(0, value.indexOf(",")),
+            "url": key.substring(5, key.length),
+            "time": value.substring(value.indexOf(",") + 2, value.length),
+          });
+        }
+      });
+
+    _myList.forEach((element) {
     });
+
+    setState(() {});
   }
 
   void mDispose() {
@@ -206,56 +214,57 @@ class _InAppBrowserState extends State<InAppBrowser>
                                           color: widget.addBookmarkIconColor,
                                         ),
                                   onTap: () async {
-                                    await _storeProvider
-                                        .checkValue(widget.mUrl)
-                                        .then((value) async {
-                                      if (value == null) {
-                                        String html = await flutterWebViewPlugin
-                                            .evalJavascript(
-                                                "window.document.getElementsByTagName('html')[0].outerHTML;");
+                                    String value =
+                                        prefs.getString("URL: ${widget.mUrl}");
 
-                                        String title = "";
-                                        if (html.contains("u003Ctitle>") &&
-                                            html.contains("u003Ctitle>"))
-                                          title = html.substring(
-                                              html.indexOf("u003Ctitle>") + 11,
-                                              html.indexOf("u003C/title>"));
-                                        else
-                                          title = "Your Url";
+                                    if (value == null) {
+                                      String html = await flutterWebViewPlugin
+                                          .evalJavascript(
+                                              "window.document.getElementsByTagName('html')[0].outerHTML;");
 
-                                        HistoryItem historyItem = HistoryItem();
+                                      String title = "";
+                                      if (html.contains("u003Ctitle>") &&
+                                          html.contains("u003Ctitle>"))
+                                        title = html.substring(
+                                            html.indexOf("u003Ctitle>") + 11,
+                                            html.indexOf("u003C/title>"));
+                                      else
+                                        title = "Your Url";
 
-                                        historyItem.title = title;
-                                        historyItem.url = widget.mUrl;
-                                        historyItem.time = DateTime.now()
-                                            .millisecondsSinceEpoch;
+                                      HistoryItem historyItem = HistoryItem();
 
-                                        _storeProvider.insert(historyItem);
+                                      historyItem.title = title;
+                                      historyItem.url = widget.mUrl;
+                                      historyItem.time =
+                                          DateTime.now().millisecondsSinceEpoch;
 
-                                        _myList.add({
-                                          "_id": historyItem.time,
-                                          "title": title,
-                                          "url": widget.mUrl,
-                                          "time": historyItem.time,
-                                        });
-                                      } else {
-                                        _storeProvider.delete(value.id);
+                                      prefs.setString(
+                                          "URL: " + historyItem.url,
+                                          historyItem.title +
+                                              ", " +
+                                              "${historyItem.time}");
 
-                                        _myList.removeWhere(
-                                          (element) =>
-                                              element["url"] == value.url,
-                                        );
-                                      }
-                                      setState(() {});
-                                    });
+                                      _myList.add({
+                                        "title": historyItem.url,
+                                        "url": historyItem.url,
+                                        "time": historyItem.time,
+                                      });
+                                    } else {
+
+                                      prefs.remove("URL: " + widget.mUrl);
+
+                                      _myList.removeWhere(
+                                        (element) =>
+                                            element["url"] == widget.mUrl,
+                                      );
+                                    }
+                                    setState(() {});
                                   },
                                 ),
                           SizedBox(width: 10),
                         ],
                       ),
                 url: widget.mUrl,
-                withLocalStorage: true,
-                withJavascript: true,
                 bottomNavigationBar: showDialog
                     ? Container(
                         height: widget.historyDialogSize,
@@ -269,15 +278,8 @@ class _InAppBrowserState extends State<InAppBrowser>
                         padding: EdgeInsets.symmetric(horizontal: 5),
                         child: Column(
                           children: [
-                            // Icon(
-                            //   Icons.remove,
-                            //   color: Colors.grey,
-                            //   size: 40,
-                            // ),
                             GestureDetector(
                               onPanUpdate: (details) {
-                                print("Mahdi: y ${details.delta.dy}");
-                                print("Mahdi: x ${details.delta.dx}");
 
                                 double appbarSize =
                                     MediaQuery.of(context).size.height;
@@ -325,8 +327,6 @@ class _InAppBrowserState extends State<InAppBrowser>
                                         ),
                                       ),
                                       onTap: () {
-                                        _myList.clear();
-                                        _storeProvider.deleteAll();
                                         setState(() {});
                                       },
                                     ),
@@ -374,11 +374,13 @@ class _InAppBrowserState extends State<InAppBrowser>
                                 itemBuilder: (ctx, index) {
                                   int dayDB =
                                       DateTime.fromMillisecondsSinceEpoch(
-                                              _myList[index]["time"])
+                                          int.parse(
+                                              "${_myList[index]["time"]}"))
                                           .day;
                                   int monthDB =
                                       DateTime.fromMillisecondsSinceEpoch(
-                                              _myList[index]["time"])
+                                          int.parse(
+                                              "${_myList[index]["time"]}"))
                                           .month;
                                   return Row(
                                     children: [
@@ -527,156 +529,3 @@ class _InAppBrowserState extends State<InAppBrowser>
           );
   }
 }
-
-// static WebViewController _webViewController;
-
-// final Completer<WebViewController> _controller =
-//     Completer<WebViewController>();
-
-// static const MethodChannel _channel = const MethodChannel('in_app_browser');
-//
-// static Future<String> get platformVersion async {
-//   final String version = await _channel.invokeMethod('getPlatformVersion');
-//   return version;
-// }
-
-// static String myUrl = "";
-// static List<String> _myList = [];
-// final flutterWebViewPlugin = FlutterWebviewPlugin();
-//
-// final Set<JavascriptChannel> jsChannels = [
-//   JavascriptChannel(
-//       name: 'Print',
-//       onMessageReceived: (JavascriptMessage message) {
-//         print(message.message);
-//       }),
-// ].toSet();
-//
-// static Widget launchUrl({
-//   @required String url,
-//   @required BuildContext contextParam,
-//   @required paddingRight,
-// }) {
-//   myUrl = url;
-//   return Scaffold(
-//     appBar: AppBar(
-//       title: Text(myUrl),
-//       leading: InkWell(
-//         child: Icon(Icons.close),
-//         onTap: () {
-//           Navigator.pop(contextParam);
-//         },
-//       ),
-//       actions: [
-//         InkWell(
-//           child: Icon(Icons.bookmark_outline_sharp),
-//           onTap: () {
-//             // _webViewController.currentUrl().then((value) {
-//             //   if (!_myList.contains(value)) {
-//             //     _myList.add(value);
-//             //   }
-//             // });
-//           },
-//         ),
-//         SizedBox(width: paddingRight),
-//       ],
-//     ),
-//     body: WebviewScaffold(
-//       url: myUrl,
-//       javascriptChannels: jsChannels,
-//       // onWebViewCreated: (WebViewController webCtrl) async {
-//       //   myUrl = await webCtrl.currentUrl();
-//       //   _webViewController = webCtrl;
-//       // },
-//     ),
-//     bottomNavigationBar: SizedBox(
-//       height: 56,
-//       child: Card(
-//         color: Colors.blue,
-//         margin: EdgeInsets.zero,
-//         shape: RoundedRectangleBorder(
-//           borderRadius: BorderRadius.circular(0),
-//         ),
-//         child: Row(
-//           mainAxisAlignment: MainAxisAlignment.spaceAround,
-//           children: [
-//             iconInkWell(
-//               func: () {
-//                 // _webViewController.canGoBack().then((value) {
-//                 //   if (value) _webViewController.goBack();
-//                 // });
-//
-//               },
-//               mIcon: Icons.arrow_back_ios,
-//               iconColor: Colors.white,
-//             ),
-//             iconInkWell(
-//               func: () {
-//                 _webViewController.canGoForward().then((value) {
-//                   if (value) _webViewController.goForward();
-//                 });
-//               },
-//               mIcon: Icons.arrow_forward_ios,
-//               iconColor: Colors.white,
-//             ),
-//             iconInkWell(
-//               func: () {
-//                 _webViewController
-//                     .currentUrl()
-//                     .then((value) => Share.share(value));
-//               },
-//               mIcon: Icons.send,
-//               iconColor: Colors.white,
-//             ),
-//             MediaQuery(
-//               data: MediaQueryData(),
-//               child: InkWell(
-//                 child: Icon(Icons.history_rounded, color: Colors.white),
-//                 onTap: () {
-//                   showModalBottomSheet(
-//                     context: contextParam,
-//                     builder: (BuildContext ctx) {
-//                       return ListView.builder(
-//                         itemCount: _myList.length,
-//                         itemBuilder: (context, index) {
-//                           return InkWell(
-//                             child: Row(
-//                               mainAxisAlignment:
-//                                   MainAxisAlignment.spaceBetween,
-//                               children: [
-//                                 Text(_myList[index]),
-//                                 Icon(Icons.arrow_forward_ios_rounded),
-//                               ],
-//                             ),
-//                             onTap: () {},
-//                           );
-//                         },
-//                       );
-//                     },
-//                   );
-//                 },
-//               ),
-//             ),
-//             iconInkWell(
-//               mIcon: Icons.refresh,
-//               iconColor: Colors.white,
-//               func: () {
-//                 _webViewController.currentUrl().then((value) {
-//                   myUrl = value;
-//                   _webViewController.reload();
-//                 });
-//               },
-//             ),
-//           ],
-//         ),
-//       ),
-//     ),
-//     // bottomNavigationBar: BottomNavigationBar(
-//     //   showSelectedLabels: false,
-//     //   showUnselectedLabels: false,
-//     //   items: [
-//     //     BottomNavigationBarItem(icon: Icon(Icons.arrow_forward_ios)),
-//     //   ],
-//     // ),
-//   );
-// }