+++ CHANGELOG.md
@@ -1,3 +1,21 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* TODO: initial release.
+
+[d8e92a7](https://github.com/AbbasJafariTrax/InAppBrowserPackage/commit/d8e92a7cacd0a1718d4aea4719bd5dd301965415),
+# Pull detection in dialog bottom sheet finished
+
+[aeb3bee](https://github.com/AbbasJafariTrax/InAppBrowserPackage/commit/aeb3bee0bbcfe3bd32bbcd97813918e204b9b7a9),
+# Bottom Sheet design and local storage setup it
+
+[6c8a6cb](https://github.com/AbbasJafariTrax/InAppBrowserPackage/commit/6c8a6cbbbe2584ef6e68b10b95b1b5aec340ec3a),
+# Dynamic Items
+
+[020e31f](https://github.com/AbbasJafariTrax/InAppBrowserPackage/commit/020e31f7de045181851d3952cbd6cab453b04170),
+# History button finished
+
+[8e4e8eb](https://github.com/AbbasJafariTrax/InAppBrowserPackage/commit/8e4e8eb36682d7700bffb969882adc1e73d9c972),
+# Share button finished
+
+[c8e09ea](https://github.com/AbbasJafariTrax/InAppBrowserPackage/commit/c8e09ea15162e458e6f1954811136655cb4ef38d),
+# Just History button and share button remining